Format the output of http://georeferencing.org/georefcalculator/gc.html

The online georeferencingcalculator allows for the results to be output by selecting 'Copy' and then pasted into a spreadsheet. This function takes the output as an argument and formats it as a tibble.

Usage

dwc_format_gco(location_identifier, georeference_calculator_output)

Arguments

location_identifier

The name of the location. Used so that the returned tibble can be joined with other data.

georeference_calculator_output

The output of the georeferencing calculator at http://georeferencing.org/georefcalculator/gc.html.

Value

A tibble. Column names are Darwin Core terms.

Examples

# Note that the georeference_calculator_output has been truncated for this
# example
dwc_format_gco(
  location_identifier = "A",
  georeference_calculator_output =
    "10.27  -123.6  WGS84  1557  0.0000001  JB  2022-08-31  Georeferencing..."
)
#> # A tibble: 1 × 9
#>   location_identifier decimalLatitude decimalLongitude geodeticDatum
#>   <chr>                         <dbl>            <dbl> <chr>        
#> 1 A                              10.3            -124. WGS84        
#> # ℹ 5 more variables: coordinateUncertaintyInMeters <dbl>,
#> #   coordinatePrecision <dbl>, georeferencedBy <chr>, gereferenceDate <chr>,
#> #   georeferenceProtocol <chr>